Initial Framework
Created a class that allows customizable parameters to be ran. Simply edit launch.py and start it to launch the program.
More parameters can be added but for now, it can pulls stocks + forex data and saves it into the directory for future usages (though I need to add the check condition if data already existed).